Few minor changes.
1. Rename EmailMockSetup to EmailExpectation.
2. Introduce ServiceProviderExtensions.DefaultPassword constant.
